+++ product_service/lib/utils/responseHelper.ts
